Wrapper function for ggmap to produce a map layer in e61 format

Simple wrapper around get_stamenmap that makes it easier to overlay a map layer for geographical data in ggplot

Usage

e61_map(
  bbox = c(top = -33.757742, right = 151.492882, bottom = -34.024779, left = 150.839539),
  adjust = 0,
  maptype = "toner-lite"
)

Arguments

bbox

A named vector of coordinates decribing the extent of your map, e.g c(top = -33, right = 151, bottom = -34, left = 150)

adjust

Adjust default zoom value

maptype

Set as the default map type for the e61 style, you can change to another maptype if needed

Details

The primary purpose of this function is to add a map layer with major place names and roads from Stamen's map server to plot under geographical data.

Examples

library(sf)
#> Error in library(sf): there is no package called ‘sf’
library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union
sa3 <- filter(absmapsdata::sa32021, gcc_code_2021 == "1GSYD") # let's just look at Sydney
#> Error in loadNamespace(x): there is no package called ‘absmapsdata’

 ggplot(sa3) +
 e61_map(bbox = c(left = min(sa3$cent_long), bottom = min(sa3$cent_lat), right = max(sa3$cent_long), top = max(sa3$cent_lat))) +
 geom_point(aes(x = cent_long, y = cent_lat)) #plot points
#> Error in ggplot(sa3): object 'sa3' not found