Wrapper function to produce e61 style graph footers

Simple wrapper around labs that makes it easier to produce good looking footer text when using footnotes and sources.

Usage

labs_e61(
  title,
  subtitle = NULL,
  footnotes = NULL,
  sources = NULL,
  title_max_char = 70L,
  subtitle_max_char = 80L,
  footnote_max_char = 85L,
  ...
)

Arguments

footnotes

A vector of footnote text strings. Each new string will be prepended with \*, \*\*, \*\*\*, etc. Please be sensible with the number of separate points you include in the graph.

sources

A vector of providing the names of sources for the graph.

title_max_char

Set the maximum number of characters per line in the title, the default is appropriate for the default graph dimensions in e61_save.

subtitle_max_char

Set the maximum number of characters per line in the subtitle, the default is appropriate for the default graph dimensions in e61_save.

footnote_max_char

Set the maximum number of characters per line in the footer, the default is appropriate for the default graph dimensions in e61_save.

...

Arguments passed on to ggplot2::labs

title

The text for the title.

subtitle

The text for the subtitle for the plot which will be displayed below the title.

caption

The text for the caption which will be displayed in the bottom-right of the plot by default.

tag

The text for the tag label which will be displayed at the top-left of the plot by default.

alt,alt_insight

Text used for the generation of alt-text for the plot. See get_alt_text for examples.

Details

The primary purpose of this function is to correctly format footer text without requiring the user to guess where to put manual line breaks for long footnotes or put in "Sources:" themselves. It does this by transforming the footnotes and sources arguments into nicely formatted text that goes into the caption argument in ggplot2's labs() function. Thus, if you are using footnotes or sources, do not supply a caption argument as well.