Format axes in the e61 Institute style

These functions format the x and y axes to be consistent with e61 styling. This includes removing white space at the beginning and end of each axis.

Usage

scale_y_continuous_e61(
  limits,
  sec_axis = dup_axis(),
  y_top = TRUE,
  expand_bottom = 0,
  expand_top = 0,
  ...
)

scale_x_continuous_e61(expand_left = 0, expand_right = 0.015, ...)

Arguments

limits

One of:

NULL to use the default scale range.

A numeric vector of length two providing the minimum and maximum limits of the scale.

A numeric vector of length three providing the limits of the scale and the increment between each axis tick, e.g. c(0, 25, 5) will set the axis to range from 0 to 25, with increments of 5 per tick.

sec_axis

Adds a secondary axis (defaults to dup_axis(), which duplicates the axis), see sec_axis for more details. Set to FALSE to hide a secondary axis.

y_top

Logical. Ensures there is space at the top of the y-axis for the axis label. Defaults to TRUE. Set to FALSE if the axis label is placed elsewhere.

expand_bottom

Numeric. Add extra space between data points and the bottom of the graph. See expansion for details.

expand_top

Numeric. Add extra space between data points and the top of the graph.

...

Arguments passed on to ggplot2::scale_y_continuous, ggplot2::scale_x_continuous

name

The name of the scale. Used as the axis or legend title. If waiver(), the default, the name of the scale is taken from the first mapping used for that aesthetic. If NULL, the legend title will be omitted.

labels

One of:

NULL for no labels

waiver() for the default labels computed by the transformation object

A character vector giving labels (must be same length as breaks)

An expression vector (must be the same length as breaks). See ?plotmath for details.

A function that takes the breaks as input and returns labels as output. Also accepts rlang lambda function notation.

expand

For position scales, a vector of range expansion constants used to add some padding around the data to ensure that they are placed some distance away from the axes. Use the convenience function expansion() to generate the values for the expand argument. The defaults are to expand the scale by 5% on each side for continuous variables, and by 0.6 units on each side for discrete variables.

oob

One of:

Function that handles limits outside of the scale limits (out of bounds). Also accepts rlang lambda function notation.

The default (scales::censor()) replaces out of bounds values with NA.

scales::squish() for squishing out of bounds values into range.

scales::squish_infinite() for squishing infinite values into range.

na.value

Missing values will be replaced with this value.

trans

For continuous scales, the name of a transformation object or the object itself. Built-in transformations include "asn", "atanh", "boxcox", "date", "exp", "hms", "identity", "log", "log10", "log1p", "log2", "logit", "modulus", "probability", "probit", "pseudo_log", "reciprocal", "reverse", "sqrt" and "time".

A transformation object bundles together a transform, its inverse, and methods for generating breaks and labels. Transformation objects are defined in the scales package, and are called <name>_trans (e.g., scales::boxcox_trans()). You can create your own transformation with scales::trans_new().

guide

A function used to create a guide or its name. See guides() for more information.

position

For position scales, The position of the axis. left or right for y axes, top or bottom for x axes.

breaks

One of:

NULL for no breaks

waiver() for the default breaks computed by the transformation object

A numeric vector of positions

A function that takes the limits as input and returns breaks as output (e.g., a function returned by scales::extended_breaks()). Also accepts rlang lambda function notation.

minor_breaks

One of:

NULL for no minor breaks

waiver() for the default breaks (one minor break between each major break)

A numeric vector of positions

A function that given the limits returns a vector of minor breaks. Also accepts rlang lambda function notation.

n.breaks

An integer guiding the number of major breaks. The algorithm may choose a slightly different number to ensure nice break labels. Will only have an effect if breaks = waiver(). Use NULL to use the default number of breaks given by the transformation.

sec.axis

sec_axis() is used to specify a secondary axis.

expand_left

Numeric. Add extra space between data points and the left of the graph.

expand_right

Numeric. Add extra space between data points and the right of the graph.