e61 themed spatial maps options

Applies the e61 theme to ggplot spatial maps to adjust graph appearance. If you are looking to change the appearance of titles or labels, check the arguments in labs_e61(), which are probably what you are looking for.

Usage

theme_e61_spatial(
  legend = c("none", "bottom", "top", "left", "right"),
  legend_title = FALSE,
  base_size = 10,
  base_family = "pt-sans"
)

Arguments

legend

Character. Legend position, "none" (default) hides the legend.

legend_title

Logical. Include Legend title? Defaults to FALSE.

base_size

Numeric. Chart font size. Default is 10.

base_family

Character. Chart font family. Default is PT Sans.

aspect_ratio

Numeric. Sets the aspect ratio of the graph panel.

Value

ggplot2 object

See also

Examples

if (FALSE) { # \dontrun{
sa3_shp <- strayr::read_absmap("sa32016")

sydney_map <- filter(sa3_shp, gcc_code_2016 == "1GSYD")

ggplot(data = sydney_map) +
  geom_sf(aes(fill = gcc_name_2016)) +
  theme_e61_spatial()
} # }